Small change to silence errors on make release
If SUPPORT isn't defined to an existing directory in xxx, make release throws error messages about not being able to find an include file when make goes to build its dependency web. However, make release is what is writing SUPPORT to be a given directory, so we'll just make the includes optional to quiet the build.